@@ -1,16 +1,17 @@
 name: awareframework_wifi
 description: This sensor allows us to handle WiFi conditions and events.
-version: 0.0.4
+version: 0.2.0
 author: AWARE Mobile Context Instrumentation Middleware Framework <yuuki.nishiyama@oulu.fi>
 homepage: http://www.awareframework.com
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  awareframework_core: ^0.1.1
+  awareframework_core: ^0.6.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
