@@ -2,8 +2,6 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 import 'package:awareframework_core/awareframework_core.dart';
-import 'package:flutter/material.dart';
-
 
 /// The WiFi measures the acceleration applied to the sensor
 /// built-in into the device, including the force of gravity.
@@ -48,18 +46,27 @@ import 'package:flutter/material.dart';
 /// var card = WiFiCard(sensor: sensor);
 /// ```
 class WiFiSensor extends AwareSensor {
-  static const MethodChannel _wifiMethod = const MethodChannel('awareframework_wifi/method');
+  static const MethodChannel _wifiMethod =
+      const MethodChannel('awareframework_wifi/method');
   // static const EventChannel  _wifiStream  = const EventChannel('awareframework_wifi/event');
 
-  static const EventChannel  _onWiFiAPDetectedStream  = const EventChannel('awareframework_wifi/event_on_wifi_ap_detected');
-  static const EventChannel  _onWiFiDisabledStream    = const EventChannel('awareframework_wifi/event_on_wifi_disabled');
-  static const EventChannel  _onWiFiScanStartedStream = const EventChannel('awareframework_wifi/event_on_wifi_scan_started');
-  static const EventChannel  _onWiFiScanEndedStream   = const EventChannel('awareframework_wifi/event_on_wifi_scan_ended');
+  static const EventChannel _onWiFiAPDetectedStream =
+      const EventChannel('awareframework_wifi/event_on_wifi_ap_detected');
+  static const EventChannel _onWiFiDisabledStream =
+      const EventChannel('awareframework_wifi/event_on_wifi_disabled');
+  static const EventChannel _onWiFiScanStartedStream =
+      const EventChannel('awareframework_wifi/event_on_wifi_scan_started');
+  static const EventChannel _onWiFiScanEndedStream =
+      const EventChannel('awareframework_wifi/event_on_wifi_scan_ended');
 
-  StreamController<WiFiScanData> onWiFiAPDetectedStreamController = StreamController<WiFiScanData>();
-  StreamController<dynamic> onWiFiDisabledStreamController = StreamController<dynamic>();
-  StreamController<dynamic> onWiFiScanStartedStreamController = StreamController<dynamic>();
-  StreamController<dynamic> onWiFiScanEndedStreamController = StreamController<dynamic>();
+  StreamController<WiFiScanData> onWiFiAPDetectedStreamController =
+      StreamController<WiFiScanData>();
+  StreamController<dynamic> onWiFiDisabledStreamController =
+      StreamController<dynamic>();
+  StreamController<dynamic> onWiFiScanStartedStreamController =
+      StreamController<dynamic>();
+  StreamController<dynamic> onWiFiScanEndedStreamController =
+      StreamController<dynamic>();
 
   WiFiScanData wiFiScanData = WiFiScanData();
 
@@ -68,7 +75,7 @@ class WiFiSensor extends AwareSensor {
   /// ```dart
   /// var sensor = WiFiSensor.init(null);
   /// ```
-  WiFiSensor():this.init(null);
+  WiFiSensor() : super(null);
 
   /// Init WiFi Sensor with WiFiSensorConfig
   ///
@@ -80,7 +87,7 @@ class WiFiSensor extends AwareSensor {
   ///
   /// var sensor = WiFiSensor.init(config);
   /// ```
-  WiFiSensor.init(WiFiSensorConfig config) : super(config){
+  WiFiSensor.init(WiFiSensorConfig config) : super(config) {
     super.setMethodChannel(_wifiMethod);
   }
 
@@ -100,7 +107,8 @@ class WiFiSensor extends AwareSensor {
     return onWiFiAPDetectedStreamController.stream;
   }
 
-  StreamController<dynamic> initStreamController(StreamController<dynamic> controller){
+  StreamController<dynamic> initStreamController(
+      StreamController<dynamic> controller) {
     controller.close();
     controller = StreamController<dynamic>();
     return controller;
@@ -120,29 +128,37 @@ class WiFiSensor extends AwareSensor {
 
   @override
   Future<Null> start() {
-    super.getBroadcastStream(_onWiFiAPDetectedStream, 'on_wifi_ap_detected').map(
-            (dynamic event) => WiFiScanData.from(Map<String,dynamic>.from(event))
-    ).listen((event){
+    super
+        .getBroadcastStream(_onWiFiAPDetectedStream, 'on_wifi_ap_detected')
+        .map((dynamic event) =>
+            WiFiScanData.from(Map<String, dynamic>.from(event)))
+        .listen((event) {
       wiFiScanData = event;
-      if (!onWiFiAPDetectedStreamController.isClosed){
+      if (!onWiFiAPDetectedStreamController.isClosed) {
         onWiFiAPDetectedStreamController.add(event);
       }
     });
 
-    super.getBroadcastStream(_onWiFiDisabledStream, 'on_wifi_disabled').listen((event){
-      if (!onWiFiDisabledStreamController.isClosed){
+    super
+        .getBroadcastStream(_onWiFiDisabledStream, 'on_wifi_disabled')
+        .listen((event) {
+      if (!onWiFiDisabledStreamController.isClosed) {
         onWiFiDisabledStreamController.add(event);
       }
     });
 
-    super.getBroadcastStream(_onWiFiScanStartedStream, 'on_wifi_scan_started').listen((event){
-      if(!onWiFiScanStartedStreamController.isClosed){
+    super
+        .getBroadcastStream(_onWiFiScanStartedStream, 'on_wifi_scan_started')
+        .listen((event) {
+      if (!onWiFiScanStartedStreamController.isClosed) {
         onWiFiScanStartedStreamController.add(event);
       }
     });
 
-    super.getBroadcastStream(_onWiFiScanEndedStream, 'on_wifi_scan_ended').listen((event){
-      if(!onWiFiScanEndedStreamController.isClosed){
+    super
+        .getBroadcastStream(_onWiFiScanEndedStream, 'on_wifi_scan_ended')
+        .listen((event) {
+      if (!onWiFiScanEndedStreamController.isClosed) {
         onWiFiScanEndedStreamController.add(event);
       }
     });
@@ -170,7 +186,7 @@ class WiFiSensor extends AwareSensor {
 ///   ..debug = true
 ///   ..frequency = 100;
 /// ```
-class WiFiSensorConfig extends AwareSensorConfig{
+class WiFiSensorConfig extends AwareSensorConfig {
   WiFiSensorConfig();
   @override
   Map<String, dynamic> toMap() {
@@ -190,107 +206,14 @@ class WiFiScanData extends AwareData {
   String security = "";
   int frequency = 0;
   int rssi = 0;
-  WiFiScanData():this.from(null);
-  WiFiScanData.from(Map<String,dynamic> data):super.from(data){
-    if (data != null){
+  WiFiScanData() : this.from({});
+  WiFiScanData.from(Map<String, dynamic>? data) : super(data ?? {}) {
+    if (data != null) {
       bssid = data["bssid"] ?? "";
       ssid = data["ssid"] ?? "";
       security = data["security"] ?? "";
       frequency = data["frequency"] ?? 0;
       rssi = data["rssi"] ?? 0;
     }
-  }
-}
-
-///
-/// A Card Widget of WiFi Sensor
-///
-/// You can generate a Cart Widget by following code.
-/// ```dart
-/// var card = WiFiCard(sensor: sensor);
-/// ```
-class WiFiCard extends StatefulWidget {
-  WiFiCard({Key key, @required this.sensor}) : super(key: key);
-
-  final WiFiSensor sensor;
-
-  @override
-  WiFiCardState createState() => new WiFiCardState();
-}
-
-///
-/// A Card State of WiFi Sensor
-///
-class WiFiCardState extends State<WiFiCard> {
-
-  String data = "AP: ---";
-  String state = "state: ---";
-
-  @override
-  void initState() {
-
-    super.initState();
-
-    if(mounted){
-      setState(() {
-        var wifi = widget.sensor.wiFiScanData;
-        data = data = "AP: ${wifi.ssid} ${wifi.bssid} ${wifi.frequency}";
-      });
-    }
-
-    // set observer
-    widget.sensor.onWiFiAPDetected.listen((wifi) {
-      if(mounted){
-        setState((){
-          if(wifi!=null){
-            DateTime.fromMicrosecondsSinceEpoch(wifi.timestamp);
-            data = "AP: ${wifi.ssid} ${wifi.bssid} ${wifi.frequency}";
-            print(data);
-          }
-        });
-      }
-
-    }, onError: (dynamic error) {
-        print('Received error: ${error.message}');
-    });
-
-    widget.sensor.onWiFiDisabled.listen((event) {
-      if(mounted){
-        setState((){
-          state = "state: disabled";
-        });
-      }
-    });
-
-    widget.sensor.onWiFiScanEnded.listen((event) {
-      if(mounted){
-        setState((){
-          state = "state: scan end";
-        });
-      }
-    });
-
-    widget.sensor.onWiFiScanStarted.listen((event) {
-      if(mounted){
-        setState((){
-          state = "state: scan start";
-        });
-      }
-    });
-
-    print(widget.sensor);
-  }
-
-
-  @override
-  Widget build(BuildContext context) {
-    return new AwareCard(
-      contentWidget: SizedBox(
-          width: MediaQuery.of(context).size.width*0.8,
-          child: new Text("$state\n$data"),
-        ),
-      title: "WiFi",
-      sensor: widget.sensor
-    );
   }
 }
